[task] refactor: move ci visibility link helpers to libs.civisibility
